Always flush pending e_steps after the last stepper isr
During deceleration with LA, as soon as the last real stepper isr is
called, force-flush all pending steps by resetting the phase, since we
know there won't be another main/LA isr pair necessary to actually
perform them.

This caused too many retraction steps to be lost in some corner cases,
mostly visible in geared extruders and very short segments (such as the
small base in a solid infill U turn).

This could be checked for earlier (when LA_phase is initialized to 0),
but simply resetting the state without repeating the check on
steps_events_completed is faster.